requirements.txt
PyYAML==6.0.2
requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
edge-tts==7.2.7
openai==1.40.0
httpx==0.27.2
gTTS==2.5.4
newspaper4k
beautifulsoup4
lxml
mutagen
pymupdf
